move getstudentsclasses from user-wrapper to class-wrapper
