+++ src/pages/dashboards/admin/AdminStudents.jsx
@@ -4,8 +4,7 @@ import { useLocation, Link } from 'wouter';
 import { useAuth } from '@clerk/clerk-react';
 import { IoPersonOutline } from "react-icons/io5";
 import { getUsers, getStudentsForExport } from '@/wrappers/user-wrapper.js';
-import { getClasses, getClassById } from '@/wrappers/class-wrapper';
-import { getStudentsClasses } from '@/wrappers/user-wrapper';
+import { getClasses, getClassById, getStudentsClasses } from '@/wrappers/class-wrapper';
 import Unauthorized from "@/pages/Unauthorized";
 import Dropdown from '@/components/Dropdown/Dropdown';
 import Button from '@/components/Button/Button';
