+++ src/wrappers/user-wrapper.js
@@ -33,17 +33,6 @@ const getUser = async (query = "") => {
   }
 };
 
-// fetch full details of all of student's classes
-const getStudentsClasses = async (studentId) => {
-  try {
-    const response = await axios.get(`/api/students-classes/${studentId}`);
-    return response.data;
-  } catch (error) {
-    console.error("Error fetching student's classes:", error);
-    throw error;
-  }
-}
-
 const updateUser = async (userId, userData) => {
   try {
     if (Object.hasOwn(userData, "firstName")) {
@@ -88,7 +77,6 @@ export {
   postUser,
   getUsers,
   getUser,
-  getStudentsClasses,
   updateUser,
   getStudentsForExport,
   deleteUser
